# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Adopcion de mascotas",
        "description": "Esta API puede usarse para obtener información sobre las categorias y mascotas, y también para crear ordenes de adopcion.",
        "version": "1.0.0"
    },
    "paths": {
        "/rest/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Obtiene una lista de todas las categorías existentes.",
                "description": "Retorna un JSON con la lista de todas las categorías existentes en la base de datos.",
                "operationId": "7df18fabfa75c3a14abd9bac59df3341",
                "responses": {
                    "200": {
                        "description": "Lista de categorías obtenida correctamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID de la categoría.",
                                                "type": "integer"
                                            },
                                            "name": {
                                                "description": "Nombre de la categoría.",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rest/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Obtener una categoría específica",
                "description": "Devuelve los detalles de una categoría en particular.",
                "operationId": "getCategoryById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría a obtener",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoría obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID de la categoría.",
                                                "type": "integer"
                                            },
                                            "name": {
                                                "description": "Nombre de la categoría.",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoría no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Categoría no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/rest/orders": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Almacenar una nueva orden",
                "description": "Almacena una nueva orden en la base de datos.",
                "operationId": "1aec7fa0f0c3126e8dd032f623b28aa4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Correo electrónico del cliente.",
                                        "type": "string"
                                    },
                                    "id_pet": {
                                        "description": "ID de la mascota que se desea adoptar.",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Orden creada con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensaje de éxito.",
                                            "type": "string"
                                        },
                                        "order_id": {
                                            "description": "ID de la orden creada.",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "La mascota no existe o ya ha sido adoptada."
                    }
                }
            }
        },
        "/rest/pets": {
            "get": {
                "tags": [
                    "Pets"
                ],
                "summary": "Obtener todas las mascotas",
                "description": "Obtiene todas las mascotas que estan disponibles para adoptar.",
                "operationId": "183e8f0af204d375d4c65b2a1911f465",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID de la mascota.",
                                                "type": "integer"
                                            },
                                            "name": {
                                                "description": "Nombre de la mascota.",
                                                "type": "string"
                                            },
                                            "image": {
                                                "description": "URL de la imagen de la mascota.",
                                                "type": "string"
                                            },
                                            "sex": {
                                                "description": "Sexo de la mascota.",
                                                "type": "string"
                                            },
                                            "category_name": {
                                                "description": "Nombre de la categoría de la mascota.",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No hay mascotas disponibles para adoptar."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "rest/pets/{id}": {
            "get": {
                "tags": [
                    "Pets"
                ],
                "summary": "Obtener una mascota por su ID",
                "description": "Obtiene una mascota específica almacenada en la base de datos según su ID.",
                "operationId": "13e2a670ae5a1011f19fc5040c0953b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la mascota a obtener",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID de la mascota.",
                                                "type": "integer"
                                            },
                                            "name": {
                                                "description": "Nombre de la mascota.",
                                                "type": "string"
                                            },
                                            "image": {
                                                "description": "URL de la imagen de la mascota.",
                                                "type": "string"
                                            },
                                            "sex": {
                                                "description": "Sexo de la mascota.",
                                                "type": "string"
                                            },
                                            "category_name": {
                                                "description": "Nombre de la categoría de la mascota.",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Mascota no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rest/pets/category/{categoryName}": {
            "get": {
                "tags": [
                    "Pets"
                ],
                "summary": "Obtener mascotas por categoría",
                "description": "Obtiene todas las mascotas de una categoría específica.",
                "operationId": "3800ae3b2e55cb3b3c3b8de1e97d7089",
                "parameters": [
                    {
                        "name": "categoryName",
                        "in": "path",
                        "description": "Nombre de la categoría",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID de la mascota.",
                                                "type": "integer"
                                            },
                                            "name": {
                                                "description": "Nombre de la mascota.",
                                                "type": "string"
                                            },
                                            "image": {
                                                "description": "URL de la imagen de la mascota.",
                                                "type": "string"
                                            },
                                            "sex": {
                                                "description": "Sexo de la mascota.",
                                                "type": "string"
                                            },
                                            "category_name": {
                                                "description": "Nombre de la categoría de la mascota.",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Categoría no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}